Rename versioning.py to version.py.
